requirements.txt
numpy>=2.0.0
opencv-python>=4.10.0.84
PyMuPDF>=1.24.0

customtkinter>=5.2.2
Pillow>=10.0.0
